Config and Callout Trigger Changes
Added compatibility with CalloutProbability when starting a random callout. This can be enabled/disabled in the settings file and a modifier can be set (higher = probability becomes more varied)
	* Callouts are now stored in the cache after settings are loaded, not when RandomCallouts.cs initializes
	* Logs are now more friendly, and won't show assembly info (such as version) anymore
Removed the intermediate config loading variables, settings are now loaded directly
Ported the LSPDFR 0.3.1 behavior of stopping the current callout if one is running instead of calling another one in the same call. Toggable in settings
Cleaned up some code